requirements.txt
nltk
pandas
scikit-learn
datasets
JPype1
numpy
matplotlib
python-dotenv